refactor(workflow-designer-ui): implement usage limit warning and hook for agent time limits
